requirements.txt
# For installing the local package in editable mode
-e .

# Dependencies for running example notebooks and development
langchain-openai>=0.1.0 # Or your specific LLM provider package and version
python-dotenv>=1.0.0
ipywidgets>=7.0.0      # For Jupyter notebook interactivity
tabulate>=0.8.0        # For pretty-printing tables, if used in notebooks
jupyterlab             # Or 'notebook' if you prefer that

# For testing (will be used in Phase 3)
pytest>=7.0.0
pytest-mock>=3.0.0

# Optional: for linting/formatting
# black
# flake8
